print log & remove ftl_test script
